Almost complete rewrite of the old API. Bug fixes not included.

+++ internal/structures/transaction.go
@@ -11,7 +11,7 @@ type Transaction struct {
 	TransactionType TransactionType `json:"transaction_type"`
 	UserId          string          `json:"user_id"`
 	ClaimId         string          `json:"claim_id"`
-	Credits         float64         `json:"credits"`
+	Credits         float32         `json:"credits"`
 	Timestamp       int64           `json:"timestamp"`
 	Description     string          `json:"description"`
 }
